[SOIN] Extrais le contenu de la carte item
...afin de conditionner prochainement le conteneur, à savoir un lien ou un bloc
